backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arbitrage-backend"
version = "0.1.0"
description = "All-in-one arbitrage orchestration backend"
authors = [{ name = "Codex Assistant" }]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "pandas>=2.2.0",
    "redis>=5.0.1",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "pydantic-settings>=2.2.1",
    "apscheduler>=3.10.4",
    "ccxt>=4.1.78",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "ruff>=0.4.4",
    "mypy>=1.9.0",
    "types-redis",
    "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true
